refactor: move Game context out of Page component
