+++ frontend/cypress/parallel-weights.json
@@ -1,1 +1,1 @@
-{"cypress/e2e/auth/forgotPassword.cy.ts":{"time":7966,"weight":7},"cypress/e2e/auth/signin.cy.ts":{"time":6823,"weight":6},"cypress/e2e/auth/signup.cy.ts":{"time":6485,"weight":6},"cypress/e2e/auth/verifyEmailPage.cy.ts":{"time":7870,"weight":7},"cypress/e2e/calendar/calendar.cy.ts":{"time":23412,"weight":21},"cypress/e2e/calendar/eventEditor.cy.ts":{"time":17050,"weight":15},"cypress/e2e/courses/view/modelGames.cy.ts":{"time":2835,"weight":2},"cypress/e2e/games/import/importCustomPosition.cy.ts":{"time":8351,"weight":7},"cypress/e2e/games/import/importOnlineGame.cy.ts":{"time":42376,"weight":39},"cypress/e2e/games/import/importPgnText.cy.ts":{"time":17775,"weight":16},"cypress/e2e/games/import/importStartingPosition.cy.ts":{"time":5126,"weight":4},"cypress/e2e/games/list/list.cy.ts":{"time":22516,"weight":20},"cypress/e2e/games/view/explorer.cy.ts":{"time":7803,"weight":7},"cypress/e2e/help/help.cy.ts":{"time":2733,"weight":2},"cypress/e2e/landing/landing.cy.ts":{"time":5256,"weight":4},"cypress/e2e/material/books.cy.ts":{"time":1813,"weight":1},"cypress/e2e/material/memorizeGames.cy.ts":{"time":7955,"weight":7},"cypress/e2e/material/modelGames.cy.ts":{"time":5573,"weight":5},"cypress/e2e/material/ratingConversions.cy.ts":{"time":1530,"weight":1},"cypress/e2e/material/sparringPositions.cy.ts":{"time":9029,"weight":8},"cypress/e2e/navbar/navbar.cy.ts":{"time":15520,"weight":14},"cypress/e2e/profile/directories/DirectoriesTab.cy.ts":{"time":19363,"weight":18},"cypress/e2e/recent/recent.cy.ts":{"time":3972,"weight":3},"cypress/e2e/scoreboard/scoreboard.cy.ts":{"time":11696,"weight":10},"cypress/e2e/scoreboard/search.cy.ts":{"time":10148,"weight":9},"cypress/e2e/scoreboard/statistics.cy.ts":{"time":4007,"weight":3},"cypress/e2e/tournaments/calendar.cy.ts":{"time":14873,"weight":13},"cypress/e2e/tournaments/info.cy.ts":{"time":2736,"weight":2},"cypress/e2e/tournaments/leaderboard.cy.ts":{"time":3061,"weight":2},"cypress/e2e/tournaments/openClassical/submitResults.cy.ts":{"time":26762,"weight":24}}
+{"cypress/e2e/auth/forgotPassword.cy.ts":{"time":7188,"weight":7},"cypress/e2e/auth/signin.cy.ts":{"time":6276,"weight":6},"cypress/e2e/auth/signup.cy.ts":{"time":5966,"weight":5},"cypress/e2e/auth/verifyEmailPage.cy.ts":{"time":7096,"weight":7},"cypress/e2e/calendar/calendar.cy.ts":{"time":22546,"weight":22},"cypress/e2e/calendar/eventEditor.cy.ts":{"time":15980,"weight":15},"cypress/e2e/courses/view/modelGames.cy.ts":{"time":2782,"weight":2},"cypress/e2e/games/import/importCustomPosition.cy.ts":{"time":7812,"weight":7},"cypress/e2e/games/import/importOnlineGame.cy.ts":{"time":40558,"weight":40},"cypress/e2e/games/import/importPgnText.cy.ts":{"time":16053,"weight":15},"cypress/e2e/games/import/importStartingPosition.cy.ts":{"time":4654,"weight":4},"cypress/e2e/games/list/list.cy.ts":{"time":21100,"weight":20},"cypress/e2e/games/view/explorer.cy.ts":{"time":7659,"weight":7},"cypress/e2e/help/help.cy.ts":{"time":2590,"weight":2},"cypress/e2e/landing/landing.cy.ts":{"time":5436,"weight":5},"cypress/e2e/material/books.cy.ts":{"time":1829,"weight":1},"cypress/e2e/material/memorizeGames.cy.ts":{"time":7503,"weight":7},"cypress/e2e/material/modelGames.cy.ts":{"time":4613,"weight":4},"cypress/e2e/material/ratingConversions.cy.ts":{"time":1269,"weight":1},"cypress/e2e/material/sparringPositions.cy.ts":{"time":8857,"weight":8},"cypress/e2e/navbar/navbar.cy.ts":{"time":14183,"weight":14},"cypress/e2e/profile/directories/DirectoriesTab.cy.ts":{"time":18022,"weight":17},"cypress/e2e/recent/recent.cy.ts":{"time":3416,"weight":3},"cypress/e2e/scoreboard/scoreboard.cy.ts":{"time":10811,"weight":10},"cypress/e2e/scoreboard/search.cy.ts":{"time":9103,"weight":9},"cypress/e2e/scoreboard/statistics.cy.ts":{"time":3666,"weight":3},"cypress/e2e/tournaments/calendar.cy.ts":{"time":14431,"weight":14},"cypress/e2e/tournaments/info.cy.ts":{"time":3000,"weight":2},"cypress/e2e/tournaments/leaderboard.cy.ts":{"time":2960,"weight":2},"cypress/e2e/tournaments/openClassical/submitResults.cy.ts":{"time":24092,"weight":23}}
