+++ frontend/src/board/pgn/PgnBoard.tsx
@@ -1,3 +1,5 @@
+import useGame from '@/context/useGame';
+import LoadingPage from '@/loading/LoadingPage';
 import { Chess, Observer } from '@jackstenglein/chess';
 import { Box } from '@mui/material';
 import { Color } from 'chessground/types';
@@ -12,8 +14,6 @@ import React, {
     useRef,
     useState,
 } from 'react';
-import { useGame } from '../../games/view/GamePage';
-import LoadingPage from '../../loading/LoadingPage';
 import { BoardApi } from '../Board';
 import ResizableContainer from './ResizableContainer';
 import { UnderboardTab } from './boardTools/underboard/Underboard';
