+++ frontend/src/board/pgn/boardTools/underboard/Underboard.tsx
@@ -1,3 +1,4 @@
+import useGame from '@/context/useGame';
 import { useLightMode } from '@/style/useLightMode';
 import {
     AccessAlarm,
@@ -20,7 +21,6 @@ import {
 import React, { forwardRef, useImperativeHandle, useState } from 'react';
 import { Resizable, ResizeCallbackData } from 'react-resizable';
 import { AuthStatus, useAuth } from '../../../../auth/Auth';
-import { useGame } from '../../../../games/view/GamePage';
 import { useChess } from '../../PgnBoard';
 import ResizeHandle from '../../ResizeHandle';
 import Explorer from '../../explorer/Explorer';
