+++ frontend/src/board/pgn/boardTools/underboard/comments/Comments.tsx
@@ -1,3 +1,7 @@
+import { useReconcile } from '@/board/Board';
+import { useChess } from '@/board/pgn/PgnBoard';
+import useGame from '@/context/useGame';
+import { Game, PositionComment } from '@/database/game';
 import { Chess, EventType, Move } from '@jackstenglein/chess';
 import {
     Button,
@@ -8,12 +12,8 @@ import {
     TextField,
     Typography,
 } from '@mui/material';
-import React, { createContext, useContext, useEffect, useState } from 'react';
+import { createContext, useContext, useEffect, useState } from 'react';
 import { useLocalStorage } from 'usehooks-ts';
-import { Game, PositionComment } from '../../../../../database/game';
-import { useGame } from '../../../../../games/view/GamePage';
-import { useReconcile } from '../../../../Board';
-import { useChess } from '../../../PgnBoard';
 import Comment from './Comment';
 import CommentEditor, { CommentEditorProps } from './CommentEditor';
