+++ frontend/src/profile/directories/DirectoriesTab.tsx
@@ -2,7 +2,7 @@ import NotFoundPage from '@/NotFoundPage';
 import { useApi } from '@/api/Api';
 import { useRequest } from '@/api/Request';
 import { useAuth } from '@/auth/Auth';
-import { useGame } from '@/games/view/GamePage';
+import useGame from '@/context/useGame';
 import { useDataGridContextMenu } from '@/hooks/useDataGridContextMenu';
 import { useSearchParams } from '@/hooks/useSearchParams';
 import LoadingPage from '@/loading/LoadingPage';
